Decouple data node and associated volumes
When updating an AMI, data nodes would be recreated (as expected);
However data volume creation was tightly coupled to the data nodes and
associated availability zones, this meant, on the destruction of a data
node, the data nodes volumes would also be destroyed.

This change decouples the data volumes from the data nodes in a way that
doesn't want to destroy any volumes when updating AMIs. This is achieved
by calculating the desired availability zone (AV) using alternate means.

* Decouple data node / data volume association.
* Calculate data volumes desired AV using the "subnet" data source.